% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createExampleData.R
\name{createExampleData}
\alias{createExampleData}
\title{createExampleData}
\usage{
createExampleData(
  samples = c("WT_1", "WT_2", "KO_1", "KO_2"),
  group = c("WT", "WT", "KO", "KO"),
  nROI = 20,
  randomMeth = TRUE
)
}
\arguments{
\item{samples}{The sample names.}

\item{group}{The sample group names.}

\item{nROI}{The number of ROIs that should be constructed.}

\item{randomMeth}{Logical indicating whether the methylation/protection values should be randomly generated.}
}
\value{
RSE object with mock data.
}
\description{
Creates an RSE object with mock NOMe-seq data.
}
\details{
Creates an RSE object with mock NOMe-seq data.
}
\examples{
createExampleData()

}
