% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprintPerc.R
\name{footprintPerc}
\alias{footprintPerc}
\title{footprintPerc}
\usage{
footprintPerc(
  footprint_counts,
  ROIgroup = rep("motif1", nrow(footprint_counts))
)
}
\arguments{
\item{footprint_counts}{A tibble containing the sample names, ROI names,
and number of fragments in each NOMe footprint pattern category. For example
the output of the stateQuant function. Rows without any footprint counts should be removed.}

\item{ROIgroup}{A vector of the same length as the number of rows in footprint_counts,
describing a group each ROI belongs too, for example,
different transcription factor motifs at the center of the ROI.}
}
\value{
A tibble where each column corresponds to a sample-footprint percentage and each row to a ROI.
}
\description{
Calculates the percentage of all reads in a ROI-sample combination in each footprint pattern.
}
\details{
Calculates the percentage of all reads in a ROI-sample combination in each footprint pattern. Then turns
the table into wide format, where each column corresponds to a sample-footprint percentage and each row to a ROI
and clusters the rows by similarity.
}
\examples{
library(tibble)
counts <- tibble(sample = c(rep("WT",10),rep("KO",10)), ROI=rep(paste0("ROI",1:10),2),
TF=floor(runif(20,min=10,max=100)),open=floor(runif(20,min=10,max=100)),
upNuc=floor(runif(20,min=10,max=100)),Nuc=floor(runif(20,min=10,max=100)),
downNuc=floor(runif(20,min=10,max=100)))
footprintPerc(counts)


}
