% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprintQuant.R
\name{footprintQuant}
\alias{footprintQuant}
\title{footprintQuant}
\usage{
footprintQuant(
  NomeMatrix,
  nr = 2,
  window_1 = c(-50, -25),
  window_2 = c(-8, 8),
  window_3 = c(25, 50)
)
}
\arguments{
\item{NomeMatrix}{A tibble where each row corresponds to a sample ROI combination,
which should contain the following columns:
nFragsAnalyzed = the number fragments (read pairs) that was available
names = the names of the ROIs that were analyzed
SampleName = the names of the samples that were analzyed
GCH_DataMatrix = for every combination of ROI and sample, a matrix where the columns
correspond to genomic positions from start to end of the ROI and the rows to fragments.
1 = protected from GpC methylation. 0 = not protected from GpC methylation.}

\item{nr}{Integer used as a cutoff to filter sample ROI combinations that have less than
(\code{nr}) fragments analyzed (nFragsAnalyzed column).}

\item{window_1}{Integer vector with two elements representing start and end positions
of the first window relative to the ROI center.}

\item{window_2}{Integer vector with two elements representing start and end positions
of the second window relative to the ROI center.}

\item{window_3}{Integer vector with two elements representing start and end positions
of the third window relative to the ROI center.}
}
\value{
A SummarizedExperiment for the number of fragments in each footprint category,
with an assay for each pattern (see below), a column for each sample, and a row for each ROI.
tf = transcription factor footprint
open = open chromatin footprint
upNuc = upstream nucleosome footprint
downNuc = downstream nucleosome footprint
Nuc = other nucleosome footprints
no Data = fragments where no clear footprint could be assigned due to lack of data (GpCs) in one of the windows
}
\description{
Count the number of reads in each NOMe state.
}
\details{
Selects 3 windows (default is -50:-25, -8:8, 25:50) around the center of the provided region of interest (ROI) and calculates the average GpC
methylation protection for a given fragment across all GpCs in each window.
If it is above 0.5 the window is deemed protected, below 0.5, unprotected Depending on the protection pattern
in all windows, a read is put into one of 5 footprint categories: tf bound (0 - 1 - 0), open chromatin (0 - 0 - 0),
upstream nucleosome (1 - 1 - 0), other nucleosome (1 - 1 - 1, 1 - 0 - 0, 0 - 0 - 1, 1 - 0 - 1), and downstream nucleosome (0 - 1 - 1).
}
\examples{
library(tibble)
NomeMatrix <- tibble(SampleName = c(rep("WT",10),rep("KO",10)),
names=rep(paste0("ROI",1:10),2),nFragsAnalyzed=rep(20,20),
GCH_DataMatrix=rep(list(matrix(sample(c(0,1),size=150*20,
replace=TRUE),ncol=150,nrow=20)),20))
footprintQuant(NomeMatrix)

}
