% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpPercHeatmap.R
\name{fpPercHeatmap}
\alias{fpPercHeatmap}
\title{fpPercHeatmap}
\usage{
fpPercHeatmap(
  footprint_percentages,
  plotcols = c("#236467", "#AA9B39", "#822B56", "#822B26", "#822B99")
)
}
\arguments{
\item{footprint_percentages}{A tibble where each column corresponds to a sample-footprint percentage and each row to a ROI,
with the rows clustered by similarity.}

\item{plotcols}{A character vector of 5 colors to be used for the heatmaps of the 5 footprint
patterns ("tf", "open", "upNuc", "Nuc", "downNuc").}
}
\value{
Heatmaps of the percentages of all reads in a ROI-sample combination in each footprint pattern.
}
\description{
Draws heatmaps of the percentages of all reads in a ROI-sample combination in each footprint pattern.
}
\details{
Draws heatmaps of the percentages of all reads in a ROI-sample combination in each
footprint pattern ("tf", "open", "upNuc", "Nuc", "downNuc"). The rows of the heatmaps are split by ROI group.
}
\examples{
library(tibble)
counts <- tibble(sample = c(rep("WT",10),rep("KO",10)), ROI=rep(paste0("ROI",1:10),2),
tf=floor(runif(20,min=10,max=100)),open=floor(runif(20,min=10,max=100)),
upNuc=floor(runif(20,min=10,max=100)),Nuc=floor(runif(20,min=10,max=100)),
downNuc=floor(runif(20,min=10,max=100)))
randPerc <- footprintPerc(counts)
fpPercHeatmap(randPerc)


}
