% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareFootprints.R
\name{compareFootprints}
\alias{compareFootprints}
\title{compareFootprints}
\usage{
compareFootprints(
  footprint_percentages,
  res,
  WTsamples = c("WT_1", "WT_2"),
  KOsamples = c("KO_1", "KO_2"),
  plotcols
)
}
\arguments{
\item{footprint_percentages}{A tibble where each column corresponds to a sample-footprint percentage and each row to a ROI,
with the rows clustered by similarity.}

\item{res}{A tibble with the results of differential fragment count testing for each ROI-footprint combination.}

\item{WTsamples}{The control sample names (two replicates) as they appear in footprint_percentages.}

\item{KOsamples}{The treatment sample names (two replicates) as they appear in footprint_percentages.}

\item{plotcols}{A character vector of colors to be used for distinguishing the ROI groups (has to be the same length as there are ROI groups).}
}
\value{
A scatter plot for each footprint pattern comparing WT and KO percentages and significance test results.
}
\description{
Compare each footprint pattern in WT and KO samples (percentages and diNOMeTest results).
}
\details{
Plots the percentages of reads in each ROI in WT versus KO samples (mean of two replicates) in each
footprint pattern ("TF", "open", "upNuc", "Nuc", "downNuc"). The color indicates the ROI group and the shape the results of the diNOMeTest.
}
\examples{
library(tibble)
NomeMatrix <- tibble(SampleName = c(rep("WT_1",5),
rep("WT_2",5),rep("KO_1",5),rep("KO_2",5)),
names=rep(paste0("ROI",1:5),4),nFragsAnalyzed=rep(20,20),
GCH_DataMatrix=rep(list(matrix(sample(c(0,1),size=150*20,
replace=TRUE),ncol=150,nrow=20)),20))
footprint_counts <- footprintQuant(NomeMatrix)
res <- diNOMeTest(footprint_counts,WTsamples = c("WT_1","WT_2"),
KOsamples = c("KO_1","KO_2"))
footprint_percentages <- footprintPerc(footprint_counts)
compareFootprints(footprint_percentages,res,plotcols="black")



}
