% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diNOMeTest.R
\name{diNOMeTest}
\alias{diNOMeTest}
\title{diNOMeTest}
\usage{
diNOMeTest(
  footprint_counts,
  WTsamples = c("WT_1", "WT_2"),
  KOsamples = c("KO_1", "KO_2"),
  FDR = 0.05,
  FC = 2
)
}
\arguments{
\item{footprint_counts}{A Summarized Experiment containing the sample names (colData), ROI names (rowData),
and number of fragments in each NOMe footprint pattern category (assays). For example
the output of the footprintQuant function.}

\item{WTsamples}{The control sample names as they appear in footprint_quantifications.}

\item{KOsamples}{The treatment sample names as they appear in footprint_quantifications.}

\item{FDR}{The FDR cutoff for a ROI - footprint combination to be called regulated in the output.}

\item{FC}{The fold change cutoff for a ROI - footprint combination to be called regulated in the output.}
}
\value{
A tibble with the results of differential fragment count testing for each ROI-footprint combination.
}
\description{
Tests for differential fragment counts for each NOMe
footprint pattern compared to all patterns in two conditions.
}
\details{
Uses edgeR's quasi-likelihood methods to conveniently test for differential proportions of each
one of 5 distinct footprints between at least two control and at least two treatment samples.
}
\examples{
library(tibble)
NomeMatrix <- tibble(SampleName = c(rep("WT_1",5),
rep("WT_2",5),rep("KO_1",5),rep("KO_2",5)),
names=rep(paste0("ROI",1:5),4),nFragsAnalyzed=rep(20,20),
GCH_DataMatrix=rep(list(matrix(sample(c(0,1),size=150*20,
replace=TRUE),ncol=150,nrow=20)),20))
patternQuant <- footprintQuant(NomeMatrix)
diNOMeTest(patternQuant,WTsamples = c("WT_1","WT_2"),
KOsamples = c("KO_1","KO_2"))

}
