% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprintPerc.R
\name{footprintPerc}
\alias{footprintPerc}
\title{footprintPerc}
\usage{
footprintPerc(
  footprint_counts,
  ROIgroup = rep("motif1", nrow(footprint_counts))
)
}
\arguments{
\item{footprint_counts}{A Summarized Experiment containing the sample names (colData), ROI names (rowData),
and number of fragments in each NOMe footprint pattern category (assays). For example
the output of the footprintQuant function.}

\item{ROIgroup}{A vector of the same length as the number of rows in footprint_counts,
describing a group each ROI belongs too, for example,
different transcription factor motifs at the center of the ROI.}
}
\value{
A tibble where each column corresponds to a sample-footprint percentage and each row to a ROI.
}
\description{
Calculates the percentage of all reads in a ROI-sample combination in each footprint pattern.
}
\details{
Calculates the percentage of all reads in a ROI-sample combination in each footprint pattern. Then turns
the table into wide format, where each column corresponds to a sample-footprint percentage and each row to a ROI
and clusters the rows by similarity.
}
\examples{
library(tibble)
NomeMatrix <- tibble(SampleName = c(rep("WT_1",5),
rep("WT_2",5),rep("KO_1",5),rep("KO_2",5)),
names=rep(paste0("ROI",1:5),4),nFragsAnalyzed=rep(20,20),
GCH_DataMatrix=rep(list(matrix(sample(c(0,1),size=150*20,
replace=TRUE),ncol=150,nrow=20)),20))
footprint_counts <- footprintQuant(NomeMatrix)
footprintPerc(footprint_counts)


}
