% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaPlots.R
\name{metaPlots}
\alias{metaPlots}
\title{metaPlots}
\usage{
metaPlots(
  NomeMatrix,
  nr = 2,
  nROI = 2,
  ROIgroup = rep("motif1", nrow(NomeMatrix))
)
}
\arguments{
\item{NomeMatrix}{A tibble where each row corresponds to a sample ROI combination,
which should contain the following columns:
nFragsAnalyzed = the number fragments (read pairs) that was available
names = the names of the ROIs that were analyzed
SampleName = the names of the samples that were analzyed
GCH_DataMatrix = for every combination of ROI and sample, a matrix where the columns
correspond to genomic positions from start to end of the ROI and the rows to fragments.
1 = protected from GpC methylation. 0 = not protected from GpC methylation.}

\item{nr}{Integer used as a cutoff to filter sample ROI combinations that have less than
(\code{nr}) fragments analyzed (nFragsAnalyzed column).}

\item{nROI}{The number of ROIs that need to have a GpC methylation measuremnet at a given
position for this position to be included in the plot.}

\item{ROIgroup}{A vector of the same length as the number of rows in NomeMatrix,
describing a group each ROI belongs too, for example,
different transcription factor motifs at the center of the ROI.}
}
\value{
A tibble with the methylation protection profiles summarized across all ROIs in a certain group.
}
\description{
Plot the summarized GpC methylation protections across selected ROIs.
}
\details{
Summarizes the GpC methylation protections across selected ROIs.
}
\examples{
library(tibble)
NomeMatrix <- tibble(SampleName = c(rep("WT",10),rep("KO",10)),
names=rep(paste0("ROI",1:10),2),nFragsAnalyzed=rep(20,20),
GCH_DataMatrix=rep(list(matrix(sample(c(0,1),size=150*20,
replace=TRUE),ncol=150,nrow=20)),20))
metaPlots(NomeMatrix)

}
